requirements.txt
httpx
websocket-client
user-agents